requirements.txt
asgiref==3.5.2
Django==4.1.3
mysqlclient==2.1.1
numpy==1.23.4
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.3.0
pyparsing==3.0.9
pytesseract==0.3.10
sqlparse==0.4.3
tzdata==2022.6
